Moved SnackbarActionButtonPlacementModeConverter to Converters (#1755)
