+++ MaterialDesignThemes.Wpf/Themes/MaterialDesignTheme.Snackbar.xaml
@@ -1,4 +1,4 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:wpf="clr-namespace:MaterialDesignThemes.Wpf"
                     xmlns:converters="clr-namespace:MaterialDesignThemes.Wpf.Converters"
@@ -11,7 +11,7 @@
 
     <converters:NullableToVisibilityConverter x:Key="NullableToVisibilityConverter" />
     <converters:MathMultipleConverter x:Key="MathMultipleConverter" Operation="Multiply" />
-    <wpf:SnackbarActionButtonPlacementModeConverter x:Key="SnackbarActionButtonPlacementModeConverter" />
+    <converters:SnackbarActionButtonPlacementModeConverter x:Key="SnackbarActionButtonPlacementModeConverter" />
 
     <Style x:Key="MaterialDesignSnackbarActionButton" TargetType="Button">
         <Setter Property="VerticalAlignment" Value="Stretch" />
